Moved contents of SpikeCore.Data project to SpikeCore. Moved contents of SpikeLite.Irc project to SpikeCore. Moved contents of SpikeLite.Messaging project to SpikeCore. Created SpikeCore.MessageBus.Foundatio and moved ContainerBuilderExtensions to it. Renamed SpikeLite.Messaging namespace to SpikeLite.MessageBus.
