--- src/SpikeCore/SpikeCore.Messages/SpikeCore.Messages/SpikeCore.Messages.csproj
@@ -1,15 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <TargetFramework>netcoreapp2.1</TargetFramework>
-    </PropertyGroup>
-
-    <ItemGroup>
-      <PackageReference Include="Microsoft.AspNetCore.Identity.EntityFrameworkCore" Version="2.1.3" />
-    </ItemGroup>
-
-    <ItemGroup>
-      <ProjectReference Include="..\..\SpikeCore.Data\SpikeCore.Data.csproj" />
-    </ItemGroup>
-
-</Project>
