+++ src/SpikeCore/SpikeCore.sln
@@ -5,15 +5,11 @@ VisualStudioVersion = 15.0.27703.2035
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SpikeCore.Web", "SpikeCore.Web\SpikeCore.Web.csproj", "{DCD98E8C-EAFC-4F5D-A68A-8039A8B62B9C}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SpikeCore.Irc", "SpikeCore.Irc\SpikeCore.Irc.csproj", "{BD53A838-72F7-48B5-93B4-A2957DF2CB6D}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SpikeCore.Irc.IrcDotNet", "SpikeCore.Irc.IrcDotNet\SpikeCore.Irc.IrcDotNet.csproj", "{C012E838-E84F-421F-9C17-8DC364F0943B}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SpikeCore.Data", "SpikeCore.Data\SpikeCore.Data.csproj", "{EC5C0A45-5C02-4A98-A7BB-7E2218882407}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SpikeCore", "SpikeCore\SpikeCore.csproj", "{FC68C5CA-7182-4724-8916-6636CEB66A59}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SpikeCore", "SpikeCore\SpikeCore.csproj", "{FC68C5CA-7182-4724-8916-6636CEB66A59}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SpikeCore.Messages", "SpikeCore.Messages\SpikeCore.Messages\SpikeCore.Messages.csproj", "{008CCA3B-EED1-4948-AE72-448A2473A30A}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SpikeCore.MessageBus.Foundatio", "SpikeCore.MessageBus.Foundatio\SpikeCore.MessageBus.Foundatio.csproj", "{27A052EE-25CC-4EAA-A1A9-A83327527268}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -25,26 +21,18 @@ Global
 		{DCD98E8C-EAFC-4F5D-A68A-8039A8B62B9C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{DCD98E8C-EAFC-4F5D-A68A-8039A8B62B9C}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{DCD98E8C-EAFC-4F5D-A68A-8039A8B62B9C}.Release|Any CPU.Build.0 = Release|Any CPU
-		{BD53A838-72F7-48B5-93B4-A2957DF2CB6D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{BD53A838-72F7-48B5-93B4-A2957DF2CB6D}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{BD53A838-72F7-48B5-93B4-A2957DF2CB6D}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{BD53A838-72F7-48B5-93B4-A2957DF2CB6D}.Release|Any CPU.Build.0 = Release|Any CPU
 		{C012E838-E84F-421F-9C17-8DC364F0943B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{C012E838-E84F-421F-9C17-8DC364F0943B}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{C012E838-E84F-421F-9C17-8DC364F0943B}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{C012E838-E84F-421F-9C17-8DC364F0943B}.Release|Any CPU.Build.0 = Release|Any CPU
-		{EC5C0A45-5C02-4A98-A7BB-7E2218882407}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{EC5C0A45-5C02-4A98-A7BB-7E2218882407}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{EC5C0A45-5C02-4A98-A7BB-7E2218882407}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{EC5C0A45-5C02-4A98-A7BB-7E2218882407}.Release|Any CPU.Build.0 = Release|Any CPU
 		{FC68C5CA-7182-4724-8916-6636CEB66A59}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{FC68C5CA-7182-4724-8916-6636CEB66A59}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{FC68C5CA-7182-4724-8916-6636CEB66A59}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{FC68C5CA-7182-4724-8916-6636CEB66A59}.Release|Any CPU.Build.0 = Release|Any CPU
-		{008CCA3B-EED1-4948-AE72-448A2473A30A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{008CCA3B-EED1-4948-AE72-448A2473A30A}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{008CCA3B-EED1-4948-AE72-448A2473A30A}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{008CCA3B-EED1-4948-AE72-448A2473A30A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{27A052EE-25CC-4EAA-A1A9-A83327527268}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{27A052EE-25CC-4EAA-A1A9-A83327527268}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{27A052EE-25CC-4EAA-A1A9-A83327527268}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{27A052EE-25CC-4EAA-A1A9-A83327527268}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
