+++ src/SpikeCore/SpikeCore/Data/Models/SpikeCoreUser.cs
