+++ src/SpikeCore/SpikeCore/Irc/ChannelMessage.cs
