+++ src/SpikeCore/SpikeCore/Irc/Configuration/BotConfig.cs
