+++ src/SpikeCore/SpikeCore/Irc/IIrcClient.cs
