+++ src/SpikeCore/SpikeCore/MessageBus/IMessageHandler.cs
@@ -1,7 +1,7 @@
 ﻿using System.Threading;
 using System.Threading.Tasks;
 
-namespace SpikeCore.Messages
+namespace SpikeCore.MessageBus
 {
     public interface IMessageHandler<TMessage>
     {
