+++ src/SpikeCore/SpikeCore/MessageBus/IrcConnectMessage.cs
@@ -1,4 +1,4 @@
-namespace SpikeCore.Messages
+namespace SpikeCore.MessageBus
 {
     public class IrcConnectMessage
     {
