+++ src/SpikeCore/SpikeCore/SpikeCore.csproj
@@ -5,15 +5,11 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\SpikeCore.Data\SpikeCore.Data.csproj" />
-    <ProjectReference Include="..\SpikeCore.Irc\SpikeCore.Irc.csproj" />
-    <ProjectReference Include="..\SpikeCore.Messages\SpikeCore.Messages\SpikeCore.Messages.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
     <PackageReference Include="Foundatio" Version="7.0.1738" />
     <PackageReference Include="Microsoft.AspNetCore.Identity" Version="2.1.3" />
     <PackageReference Include="Microsoft.AspNetCore.Identity.EntityFrameworkCore" Version="2.1.3" />
+    <PackageReference Include="Microsoft.EntityFrameworkCore" Version="2.1.3" />
+    <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="2.1.3" />
     <PackageReference Include="Microsoft.Extensions.Configuration.Binder" Version="2.1.1" />
   </ItemGroup>
 </Project>
